+++ macos/Classes/MacOSWindowUtilsPlugin.swift
@@ -35,6 +35,7 @@ public class MacOSWindowUtilsPlugin: NSObject, FlutterPlugin {
             if (enableWindowDelegate) {
                 MainFlutterWindowManipulator.createFlutterWindowDelegate(methodChannel: nsWindowDelegateChannel)
             }
+            MainFlutterWindowManipulator.reset()
             result(true)
             break
             
@@ -510,6 +511,25 @@ public class MacOSWindowUtilsPlugin: NSObject, FlutterPlugin {
             
         case "performClose":
             MainFlutterWindowManipulator.performClose()
+            result(true)
+            break
+            
+        case "updateToolbarPassthroughView":
+            let id = args["id"] as! String
+            let x = args["x"] as! CGFloat
+            let y = args["y"] as! CGFloat
+            let width = args["width"] as! CGFloat
+            let height = args["height"] as! CGFloat
+            let enableDebugLayers = args["enableDebugLayers"] as! Bool
+            
+            MainFlutterWindowManipulator.updateToolbarPassthroughView(id: id, x: x, y: y, width: width, height: height, enableDebugLayers: enableDebugLayers)
+            result(true)
+            break
+            
+        case "removeToolbarPassthroughView":
+            let id = args["id"] as! String
+            
+            MainFlutterWindowManipulator.removeToolbarPassthroughView(id: id)
             result(true)
             break